+++ .eslintrc.js
@@ -51,6 +51,7 @@ module.exports = {
             }
         ],
         "@typescript-eslint/no-empty-function": "error",
+        "@typescript-eslint/no-non-null-assertion": "off",
         "@typescript-eslint/no-empty-interface": "error",
         "@typescript-eslint/no-explicit-any": "off",
         "@typescript-eslint/no-this-alias": "off",